# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FreshMood Restaurant QR Ordering API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all menu categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Menu Items
  /menu-items:
    get:
      operationId: listMenuItems
      tags: [menu-items]
      summary: List menu items (optionally filtered by category)
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu-items]
      summary: Create a menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu-items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu-items]
      summary: Get a single menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    patch:
      operationId: updateMenuItem
      tags: [menu-items]
      summary: Update a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu-items]
      summary: Delete a menu item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Tables
...
  /admin/cashiers:
    get:
      operationId: listCashiers
      tags: [admin]
      summary: List all cashiers
      responses:
        "200":
          description: List of cashiers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cashier"
    post:
      operationId: createCashier
      tags: [admin]
      summary: Create a new cashier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CashierInput"
      responses:
        "200":
          description: Created cashier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cashier"
...
  /admin/cashiers/activate:
    post:
      operationId: activateCashier
      tags: [admin]
      summary: Set or clear the active cashier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateCashierInput"
      responses:
        "200":
          description: New active cashier or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cashier"
  /admin/cashiers/{id}:
    delete:
      operationId: deleteCashier
      tags: [admin]
      summary: Delete a cashier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, name, sortOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        sortOrder:
          type: integer
    CategoryPatch:
      type: object
      properties:
        name:
          type: string
        sortOrder:
          type: integer
    MenuItem:
      type: object
      required: [id, categoryId, name, price, available, isBestSeller]
      properties:
        id:
          type: integer
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        available:
          type: boolean
        isBestSeller:
          type: boolean
    MenuItemInput:
      type: object
      required: [categoryId, name, price]
      properties:
        categoryId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        available:
          type: boolean
        isBestSeller:
          type: boolean
    MenuItemPatch:
      type: object
      properties:
        categoryId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        available:
          type: boolean
        isBestSeller:
          type: boolean
...
    Cashier:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    CashierInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
    ActivateCashierInput:
      type: object
      properties:
        cashierId:
          type: ["integer", "null"]
...